fix patch generation when there are aliases
